refactor link button and fix
