+++ src/components/Boosts/Modals/Hold.tsx
@@ -6,7 +6,7 @@ import { generateBuyUrl, generateSwapUrl } from "../../../helpers/links";
 import { Chip } from "../../Chip";
 import { H3 } from "../../H3";
 import { Label } from "../../Label";
-import { LinkButton } from "../../LinkButton";
+import LinkButton from "../../LinkButton";
 import { ModalWrapper } from "../styled";
 
 interface Props {
