+++ src/components/Boosts/Modals/ScratcherLunar.tsx
@@ -3,7 +3,7 @@ import React, { FC } from "react";
 import { useTrackedState } from "../../../context/store";
 import { generateScratcherUrl } from "../../../helpers/links";
 import { H3 } from "../../H3";
-import { LinkButton } from "../../LinkButton";
+import LinkButton from "../../LinkButton";
 import { Text } from "../../Text";
 import WarningChip from "../../WarningChip";
 import { Description, ModalWrapper } from "../styled";
