+++ src/components/Campaigns/Before.tsx
@@ -3,7 +3,7 @@ import React, { FC } from "react";
 import { CampaignInfo } from "../../hooks/useCampaignInfo";
 import { H3 } from "../H3";
 import { Label } from "../Label";
-import { LinkButton } from "../LinkButton";
+import LinkButton from "../LinkButton";
 import { Text } from "../Text";
 import { BeforeCampaignJson } from "./types";
